Add ability to register streams other than WebRTC to the signaling server
