+++ src/main/java/io/antmedia/websocket/WebSocketConstants.java
@@ -68,7 +68,9 @@ public class WebSocketConstants {
 
 	public static final String REGISTER_ORIGIN_SERVER = "registerOriginServer";
 
-	public static final String REGISTER_EDGE_SERVER = "register_edge_server";
+	public static final String REGISTER_EDGE_SERVER = "registerEdgeServer";
+
+	public static final String REGISTER_BROADCAST = "registerBroadcast";
 
 	public static final String NO_STREAM_EXIST = "no_stream_exist";
